Revert to using public DNS and cert
We will be using the public DNS and cert for the ALB to avoid the overhead and complexity of the VPC link. This push reverts the changes to use the ALB DNS and VPC link since we won't be using that logic anymore. I updated comments for what will need to be changed