[Issue #5858] Handle empty response (#5909)
## Summary

Fixes #5858 

## Changes proposed

Add `_normalize_soap_dict_for_comparison` to normalize a SOAP dictionary
for comparison.
Apply this method where necessary, add tests.

## Context for reviewers

If an empty response is returned for the list of opportunities,
grants.gov returns nothing, but we have an empty list, so the comparison
fails.

This can be seen by setting up your local API to proxy a request to
training and passing it an opportunity number that doesn't exist (a
404-like case). The comparison between our response and grants.gov's
doesn't match. I'm assuming this is because we setup an "empty" response
differently. We should fix them to match

## Validation steps

See new unit tests.