[Issue #5859] Small fixes/adjustments to sam.gov processing (#5938)
## Summary
Fixes for #5859

## Changes proposed
* Add a little more logging to extract file download to say which file
it is downloading
* Fix logic for setting the processing status of a sam.gov extract so it
actually gets set

## Context for reviewers
Processing status wasn't getting set when run because we fetch the
extract files in one transaction, and update it in a different one, so
they aren't associated with the DB session in the latter case so the
update doesn't do anything. Adding them to the DB session fixes this.