[Issue #5918]  Search Filtering - Update Opportunities search Index to use review status (#5988)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #5918  

## Changes proposed

- Modified `load_opportunities_to_index.py` to filter out opportunities
that exist in the `ExcludedOpportunityReview` table
- Added test coverage

## Context for reviewers

We have a new table `excluded_opportunity_review` that signifies all
opportunities that are in a review state and should not be available in
search results.

## Validation steps

To run these specific tests 

```bash
make test args="-vv tests/src/search/backend/test_load_opportunities_to_index.py"
```