[Issue #5958] Sort the application forms when fetching an application (#6015)
## Summary
Fixes #5958 

## Changes proposed
Modified fetching an application to sort the application forms list by
form name

## Context for reviewers
The order of app forms was random and could shuffle upon refreshing the
application page which was confusing/annoying.

Ideally the right way to do this sorting is add `order_by` to the
relationship, but we don't have the form name in that table and it
doesn't let us do joins there. I couldn't find a way to sort a
relationship during a query, so just do it in code on the DB record
which should be fine.

## Validation steps
Going to the local apply page, we now consistently get a sorted order:

<img width="736" height="837" alt="Screenshot 2025-08-20 at 3 38 02 PM"
src="https://github.com/user-attachments/assets/05056ee2-ee3e-4694-8164-a58ca0b3df9d"
/>