[Unticketed] Create a utility script for uploading forms to a given environment (#5994)
## Summary

## Changes proposed
Create a utility for uploading a form to an environment from the local
DB

## Context for reviewers
This is an idea I've had for a while to help setup forms, there are more
ideas that I want to add (especially around the instruction file), but
just want something that makes the setup we have to do in the next few
days a bit simpler.

## Validation steps
I used this to update all of the forms in dev/staging.