[Issue #5832] correctly indicate failed validations for nested required fields (#6070)
Validation for nested required fields is tricky, as the filed indicated in warnings that come back usually refer to a parent path in the schema rather than the missing field itself. This change addresses that complexity so that the UI can properly indicate which fields are failing validation.

Note that this change will add red borders to indicate nested unconditionally required fields that are failing validation. It will not:
* update links in the list of validation warnings to properly link to the field, these are still broken
* correctly identify the field in the validation warning message
* style conditionally required fields