[Issue #5991] attachment filenames after save (#6077)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #5991 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->

Adds the ability for the file upload widgets to know about the list of attachments on the application, in order to display filenames of previously uploaded files.

* a new context used to pass the list of attachments down to the widgets without prop drilling
* refactored "buildFormTree" function into a React Component
* refactored "buildField"'s rendering functionality to render components rather than making function calls

The last two pieces here were necessary to get the widget components to
properly identify the parent context in order to read its value