[Issue #6097]: Delinquent Federal Debt Radio button (#6132)
## Summary

This is only a partial fix to: Delinquent Federal Debt
[#6097](https://github.com/HHS/simpler-grants-gov/issues/6097)

## Changes proposed

I've added the Radio widget manually to the ui schema
it now displays, the YES / NO options
I tried to make it as type safe as possible
added some utils and helpers to work with the options that are
returned/created

## Context for reviewers

This does not contain the validation portion. That will come in a second
PR-

## Validation steps

1. to run locally, you would need to `run npm run build && npm start`
2. `make remake-backend && docker compose up`
3. navigate to `search`
4. search for `pilot`
5. you should see an opportunity called `Local Pilot-equivalent
Opportunity`
6. sign in
7. start new application
8. navigate to form table of the application
9. click Application for Federal Assistance (SF-424)
10. scroll down to Delinquent Federal Debt
11. when making no selection, and saving, you should not see any change
to the radio and a validation error
12. when making a selection, for example, Yes, and saving, the value
should remain Yes. same with No
13. Keep in mind validation is coming in a separate PR