[Issue #5916] sync opportunity review status task   (#6128)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #5916  

## Changes proposed

A task that handles truncating the excluded_opportunity_review table and
add all the opportunities in review status by filtering the foreign
v_opportunity_summary table.

## Context for reviewers
This task will setup the table that is queried downstream to filter out
the opportunities listed here. Note: This task will run as the last step
in the load_transform job, failure that occurs here shouldn't affect the
other tasks.

## Validation steps

Test added