[Issue #5975] Enable the Sam.gov daily job (#6163)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #5975 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Switch job flag from DISABLED to ENABLED.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Previously we built the TF for the task, but the job wasn't ready to be
enabled. This enables the task.

## Validation steps

Validate the task runs after merge.