[Issue #5917] Sync suppressed emails schedule (#6177)
## Summary

Fixes  #5917  

## Changes proposed

Add the sync suppressed job to load-transform scheduler to run in lower
envs

## Context for reviewers

In the lower env, after confirming that data has been properly loaded
into the legacy table and successfully running the
SyncOpportunityReviewStatus job manually, we verified that opportunities
in "Review" status no longer appear in search results. It is now safe to
enable the scheduled execution of this job.