Infra 6106 - Attach additional ALB certs and tidy up gateway (#6107)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6106

## Changes proposed

- Logic for optionally attaching extra certs to ALB
- Enable API gateway logging
- Point API gateway to secondary domain on ALB
- Overload certain endpoints and paths to not require API token

## Context for reviewers

This PR has ~~two~~ three main parts to it, adding logic for optionally
attaching secondary certs to the ALB, and tidying up the API gateway
setup.

### ALB changes

For this part, I wanted to setup the ALB changes to be modular enough
that it won't break in case we pass no extra domains to it, but allow it
to take in more than a single extra domain in case more are added in the
future. This is done in preparation for the cutover for the public DNS
for the api.*.simpler... to point from the ALB to the API gateway. The
DNS is already setup for alb.*.simpler..., so we can point the API
gateway to that, and be ready for the cutover when it happens

### API Gateway config changes

This part has a change for enabling logging, and a change to the logic
for what domain the proxy forwards to. For the logging change, there was
an issue from my first PR where I thought the accounts playbook would be
deployed via terraform, but that didn't happen. I merged a second PR
that disabled it until I got accounts deployed, which needed to be done
manually and is now set. Now we can enable it again and have it log as
expected. The other change is to point the gateway proxy from the main
domain, `api...`, to the secondary, `alb...`, so that the cutover is
only required to be done on the DNS side. This will let us test the
cutover before going live

There might be another change to enable the domain association, which is
currently commented out, but I'll have to test to make sure it won't
break anything first. I'll edit the PR body if I can enable it with a
blurb about it

### API Gateway endpoint changes

While testing with Matt and Brandon, we found that there are certain
endpoints that we want to not have API tokens required. To handle this,
we have to define each of those endpoints (or proxy paths) and disable
the token required option. It isn't great, but it lets us keep
everything locked down by default unless we need it

## Validation steps

Validation will be done manually. I'll call out in Slack before testing,
then deploy the changes and test. Once things are validated, I'll update
the PR and revert the changes so that I don't introduce drift