[Unticketed] Fix values to ensure more deterministic behvior (#6301)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes issue of test failures for suspected nondeterministic factory
behavior.

## Changes proposed

Fix factory values to have a more consistent SOAP payload in tests.

## Context for reviewers

Another build failure was reported with a new issue:
https://github.com/HHS/simpler-grants-gov/actions/runs/17503840148/job/49722918675#step:8:216
This is difficult to reproduce locally, but fairly sure the root cause
is a factory populating an `agency_record` value which results in new
XML entries in the SOAP payload tests.

## Validation steps

See passing unit tests post deploy.