[Issue #6209] Add support for math summation and arrays in pre/post population logic (#6300)
## Summary

Fixes #6209

## Changes proposed
* Add support for math summation of fields
* Add support for pre/post populating fields with arrays
* Add support for pre/post populating fields while fetching values from
arrays
* Allow a way to specify an order to JSON rules

## Context for reviewers
This started as a prototype / tech spec and evolved a bit into the
actual implementation. This is a fairly big refactor of how we do
pre/post population, but doesn't require reconfiguring any of our
existing configurations as the config itself just has some new features.

Summing several fields and storing it in another field requires two
things:
* A way to fetch fields including supporting arrays and relative paths
* A way to handle processing fields within arrays

Arrays were by far the most complex part of this (fetching values is
mostly done with the jsonpath_ng library). When building the rules we
need to handle running rules a variable number of times depending on the
data, and populating data required a complete rewrite of the field
population logic (note: jsonpath_ng's populate functions couldn't be
used as the way they work with arrays would cause issues - if an array
doesn't exist it makes it an object which doesn't make sense).

See the updated readme for more info on using this, as well as the
comments throughout giving plenty of examples.

## Validation steps
A considerable number of new tests were added to test every piece of
this change, this was largely built by first setting up the tests and
adjusting the code until they worked.