[Issue #6372] remove widget rendering from apply utils (#5963)
The big change here is replacing the "buildField" function, which rendered a widget for a form field, with the "getFieldConfig" function, which will return a widget type and set of props to use to render the widget. This allows us to render the form separately from building out the shape of the form, and makes testing the utility file much easier.