Temp fully proxy API Gateway (#6378)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #ISSUE 

## Changes proposed

- Temp comment out full proxy requirement for API token
- Temp comment out other endpoints since they are redundant right now

## Context for reviewers

This PR will comment out the majority of the logic for the API gateway
so that it acts like a full proxy to the underlying ECS service. This
will allow us to change the DNS from the ALB to the gateway, but not
turn the gateway on, so that we can test the functionality and have a
fast way to revert back in case any issues arise. Otherwise, we'd need
to change the DNS, which can take time due to it being a long process

## Validation steps

Validation would be to hit the API gateway URL and see if things work as
expected. I'm tempted to deploy the proxy as the none stage so we can
test fully functionality before the dev DNS change occurs