[Unticketed] Fix issue with sam.gov extracts flagging expected scenario (#6390)
<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->

## Changes proposed
When a sam.gov extract row is marked as deactivated and has an EFT
indicator, don't expect to see the primary non-eft indicator row

## Context for reviewers
The very short version of the EFT indicator logic is that we can receive
multiple rows for a given UEI from sam.gov, but anything that has the
EFT indicator set is duplicate data as far as we're concerned (it's used
to add more bank account info to an entity, something we don't look at).
We only want the primary record, but added logic to flag if we saw any
non-primary records without the primary. This fixes a case where we can
get a non-primary record without the primary (deleting one of those EFT
indicator records). This doesn't adjust any meaningful processing logic,
just our alert to ourselves that the data was in an odd state.