[Issue #6208] Create a utility for generating test opportunities (#6368)
## Summary
Fixes #6208

## Changes proposed
Created a script that fetches an opportunity from any environment, and
outputs SQL to create an opportunity like it.

## Context for reviewers
We've had an increased need to setup test opportunities, but up until
now that process required manually building SQL, and an opportunity is a
lot of data across a lot of tables and the process is incredibly
tedious, error-prone and all around not fun. This script helps automate
most of that by using an existing opportunity as a baseline, swapping
out the primary keys, and printing out the SQL you'll need to do this.

It does not yet handle copying opportunity attachment or competition
instruction files as that'll require converting the s3 CDN paths back to
s3 paths, and also generating s3 copy commands. Doable, but not that
urgent to handle, can follow-up on that when needed.

The tests aren't too elaborate, and I don't validate the SQL generated -
this is meant for helping setup test data, setting up a bunch of tests
for a utility isn't something I think is valuable time (and would take a
lot of time and be very complex).

## Validation steps
For example, let's say I wanted to copy this opportunity:
https://staging.simpler.grants.gov/opportunity/dab5a586-69e3-4710-91d5-9dfcc40ae9d4

I could run `make cmd args="task generate-opportunity-sql
--environment=staging
--opportunity-id=dab5a586-69e3-4710-91d5-9dfcc40ae9d4"` and it would
output a bunch of SQL statements to make the opportunity (this
screenshot isn't nearly all of it):
<img width="1455" height="873" alt="Screenshot 2025-09-10 at 3 17 35 PM"
src="https://github.com/user-attachments/assets/8efdd1a6-cd03-40f3-b345-0d971076f3cf"
/>

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>