[Issue #6404] Fix prepopulation when value is null (#6425)
## Summary
Fixes  #6404

## Changes proposed
Adjust behavior of prepopulation when a value is null, by default the
field should not be populated AND any existing value is removed.

Adjust some of our prepopulation logic to not let a field be null
(opportunity number, opportunity title, and agency name) to avoid issues
where a required field can't be populated

Add some additional tests to the SF424 validation pre-population
behavior

## Context for reviewers
Some prepopulation fields can be null as the field is not required, for
example the competition title is a commonly nullable field. But our JSON
schema doesn't like those fields being null, even if they're not
required. Null != missing, missing is fine, null would require we change
the type from `string` to `["string", "null"]` which is a much bigger
change. Instead we'll change the default to not include the field.
Additionally, it will delete an existing value, this is because an API
user could pass in any value for it, and we want to make sure it's unset
(there are a few very specific edge cases still unhandled, but not
needed any time soon).

For the fields I made not be able to prepopulate to null, they do have
nullable columns but in prod opportunity number and title are never
actually null. Agency is only null on about 40 opportunities, most of
those from before 2010. We should eventually fix the schema to not let
those fields be null, but that's not something we should worry about
yet.