[Issue #6352] Get operation name without parsing whole xml (#6370)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Work for #ISSUE 6352

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
We need to adjust how we determine the operation to not parse the entire
attachment -- the information we need is appearing very early in the
parsing.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->