[Issue #6430] Lowercase emails for sam.gov and login (#6432)
## Summary
Fixes #6430

## Changes proposed
When loading sam.gov data, lowercase the email.
When receiving login.gov data, lowercase the email.

When joining sam.gov data for creating orgs, skip all blank emails

## Context for reviewers
The sam.gov extract doesn't store emails in a consistent case, we might
receive an email of `ABC@mail.com` but login.gov looks like it always
lowercases it to `abc@mail.com`. This means the join wouldn't actually
work, and we're probably missing some orgs. This fixes the emails when
we import them.

Also we have a few hundred sam.gov entities without an email, just in
case, we don't want that to join with a blank email in the external
table.