[Issue #6359] Skip sending to grants.gov if simpler formatted id (#6434)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->

If it is Simpler format, skip sending request to Grants.gov

## Changes proposed

Added a check to see if the incoming request meets the following
criteria:
- It is either GetApplicationZip or GetApplication
- It has a GrantsGovTrackingNumber that starts with an 8 or 9

If both those critera are met, do not forward the request on to
grants.gov.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Hitting grants.gov can be slow, so if we know for sure the data is not
in grants.gov we won't bother actually making the request.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->