Frontend API Gateway key and secret (#6465)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #ISSUE 

## Changes proposed

- API Gateway key for frontend
- SSM param for frontend to store key in
- Update frontend ECS task def

## Context for reviewers

This PR will create an API key for the frontend to use, and a SSM param
that will require someone to copy the key value into (terraform will not
do this automatically). I setup the param in the frontend playbook
because it would cause a race condition if it was made in the API
gateway file, since frontend and API would be deployed at the same time

## Validation steps

Dependent on code and steps from @btabaska to be tested via frontend.
The gateway has been set to full proxy mode, so I would need to set an
endpoint to have auth required and then test via that path