[Issue #6516] Add retries to the sam.gov library (#6541)
## Summary
Fixes  #6516

## Changes proposed
Add retry logic to the sam.gov logic to fetch extracts as we've noticed
intermittent timeouts.

## Context for reviewers
Decided to use the tenacity library instead of the urllib3 retry as it
seemed a bit less verbose.

Removed an unused request function from the client and simplified some
of the file logic in the tests while I was adjusting things.

## Validation steps
I verified this worked with the existing test that threw a timeout
exception. Because of the new retries that'll take place over a minute
or two, that test wouldn't make sense to run anymore, but I kept it and
marked it as skipped.