[Issue #5874] Update cpu and mem config for ecs task definition (#6545)
## Summary

Updated the memory and cpu config is to be relative to the total vcpu
and memory profile of the Fargate instance.

## Context for reviewers

Currently, the full vCPU and memory of the app task definition is not
being utilized.

## Validation steps

After deploying the changes in the lower environment we should inspect
the task definition to check the full cpu and mem is utilized in the
task definition.

---------

Co-authored-by: prasnava <parikatla@Prasanthreddys-MacBook-Pro-2.local>