[Issue #6520] List available organization roles (#6556)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #[6342](https://github.com/HHS/simpler-grants-gov/issues/6342)

## Changes propose
`POST` endpoint for returning organization roles given the user has the
`view_org_membership` privilege on the requested organization.
For the time being we only have core roles and those are being returned.
Schema for the response
Test added

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>