[Issue #6360] Create process for testing soap endpoints (#6535)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #ISSUE 6360

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Create prototype for testing SOAP endpoints and validating schemas

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Run the same request against the proxy and the S2S API directly and
confirm the responses are identical

This can potentially be done in Postman or a Script.
Documenting this can become a model for something going forward.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->