[Issue #6494] Update migration for `form_type_id` foreign key [2/2] (#6586)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6494 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Update migration to account for foreign key relationship.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
There was an error running the migration on the merge of the first PR
for this ticket. This fixes the issue.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
run `make db-migrate` successfully.