[Issue #6397] Update schemas / scripts include json_to_xml_schema (#6611)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6397 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Include `json_to_xml_schema` in form schemas (GET and PUT)
Add `json_to_xml_schema` to diff checking
Update tests

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
We want to update the PUT form endpoint to include the new
json_to_xml_schema in its API schema. This value should also be used to
update the form record in the DB like all of the other fields. Note that
because the endpoint in the environments won't actually have the first
change yet, the update won't "work" yet.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See tests for GET/PUT updates.
Due to the diff checking working on the keys of the `dict`, we just need
to add `json_to_xml_schema` for checking, and the difference should be
picked up.