[Issue #6585] Stream soap proxy response data instead of chunking (#6592)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #ISSUE6585 

## Changes proposed

Modify SoapResponse to take in bytes or an iterator for data. This is so
we can stream data from the soap proxy instead of chunking it into a
string.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Currently modifying SoapResponse to perform identically if you pass it
an iterator or bytes. Adding the iterator did add some challenges
because that iterator could be exhausted, so I modified SoapResponse to
cache that data. I also put in a boolean check for the part of the
client that compares the two responses (and logs the difference) since
you have to stop streaming in order to do that. In the future if we want
to, we can flip the boolean based on the operation_name.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->