[Issue #6490] Modify update-form script to not need form instruction ID (#6564)
## Summary
Fixes #6490

## Changes proposed
* On each statically defined form, add a form instruction ID
* Adjust the seed script to make a form instruction record for anything
with a form instruction ID
* Adjust the form scripts to expect form instruction ID on the form we
have locally

## Context for reviewers
Because form instructions work a little different than other fields on
the form, needing the ID to point to a different record you have to
manually setup, I had made it so the script required you pass it in. But
that's really not needed, so made it so we define the ID on the forms
themselves and just use that when diffing/updating forms. Most of the
logic changes are just simplifying that.

Note that in the db-seed-local changes, I made it so it generates txt
files instead of PDFs. Making a PDF with a little bit of content is way
too complex for a seed script. Surprisingly there isn't just a way to
say "make a PDF with a little text" that isn't a 5 step process or
requires several other PDF libraries.

## Validation steps
Checking each environment, these changes (as expected) don't show any
changes to the forms as we already had the form instruction ID set on
all of them.

Had to do a bit of adjusting to make sure `make db-seed-local` would
setup the instruction forms as expected, but the script works when
re-running it.

The output of `make cmd args="task list-forms --environment=prod
--verbose"`
```
+------------------------------------------------------------+---------+-------------------------+----------------+
| Form Name                                                  | Version | Last Updated            | Changed Fields |
+------------------------------------------------------------+---------+-------------------------+----------------+
| Assurances for Non-Construction Programs (SF-424B)         | 1.1     | 2025-08-21 16:10:27 UTC |                |
| Project Abstract Summary                                   | 2.0     | 2025-08-21 16:10:41 UTC |                |
| Project Narrative Attachment Form                          | 1.2     | 2025-08-21 16:10:47 UTC |                |
| Budget Narrative Attachment Form                           | 1.2     | 2025-08-21 16:10:53 UTC |                |
| Application for Federal Assistance (SF-424)                | 4.0     | 2025-09-24 16:35:21 UTC |                |
| Budget Information for Non-Construction Programs (SF-424A) | 1.0     | 2025-09-24 16:35:38 UTC |                |
| Disclosure of Lobbying Activities (SF-LLL)                 | 2.0     | 2025-09-24 16:35:49 UTC |                |
+------------------------------------------------------------+---------+-------------------------+----------------+
```