[Issue #6603] Add ebiz_poc_email, ebiz_poc_first_name, ebiz_poc_last_name (#6702)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6603 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
- Added `ebiz_poc_email`, `ebiz_poc_first_name`, and
`ebiz_poc_last_name` fields to `SamGovEntityResponseSchema`.
- Updated tests

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
The organization info fields existed in the `SamGovEntity` database
model, but they were not being exposed in the API response schema. This
adds these fields to the `SamGovEntityResponseSchema` so they are
included in org info API responses.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See updated tests

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>