Add API root proxying for Gateway (#6704)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #ISSUE 

## Changes proposed

- Add routing logic for the root path
- Add comments for future PR

## Context for reviewers

This PR fixes a finding Michael had from the gateway not having any
routing logic for the root path. This PR adds that, and some comments
for future work on what needs to have non-API token required logic

## Validation steps

This was done manually for a few minutes in dev, and both Michael and I
confirmed the page was loading as expected. Before merging this, I'll
make sure to deploy and test to make sure. AWS console and terraform
have different restrictions