[Issue #6665] Define schemas for GetApplicationZip (#6721)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #ISSUE6665

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Added schemas for request and response for GetApplicationZip.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
This is to prove we can consume request xml and convert it to a dict and
take a response dict and convert it to xml, with some additional MTOM
header.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->