[Issue #6747] Add submitted_at and submitted_by to Application table (#6747)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6655 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Adds `submitted_at` and `submitted_by` to the application table,
including a generated migration.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
As part of the work to generate XML to meet existing Grants.gov needs,
we need to store when an application was submitted and who submitted it.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
Run migration and confirm new columns are in the database.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>