[Issue #6689] Add USE_SIMPLER flag to env variables (#6729)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #ISSUE6689 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Added USE_SIMPLER variable to the environment and set them to True in
lower environments and False in prod. This should allow us to eventually
start testing.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Prior to this the variable `use_simpler_response` was hardcoded to False
so we were never getting a simpler response.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->