[Issue #6588] Remove not-started application form status (#6745)
## Summary
Work for #6588

## Changes proposed
Remove application form status of `not_started` and logic that would set
it.

## Context for reviewers
This status was causing issues with the frontend and was calculated in a
very obtuse way. This basically merges what was `not_started` into the
`in_progress` status.

The frontend already only uses `in_progress` so this fixes some issues
for them.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>