[Issue #6488] Update PUT and GET endpoint to include new sgg_version / is_deprecated / form_type columns (#6781)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6488

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Add sgg_version / is_deprecated / form_type to `FormAlphaSchema` and
`FormUpdateRequestSchema`, which is responsible for `GET` and `PUT`
endpoints.
Add same fields to `form_task_shared.py]`
Update tests

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Adds fields to the endpoints and the script that talks with the
endpoints. In support of creating a process for deploying forms to
production environment.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See unit tests

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>